Improved patterns
Rizon is terrible and decided to require RECOVER and say we can't use
GHOST. So I rehauled the pattern matching... it needed it anyway. Also
increased the timeout between ghost attempts. Also keeps it from trying
to set the nick until a message comes in saying the nick is free (avoids
a nasty dialog in pidgin)